+++ SharpMC/Items/ItemFactory.cs
@@ -2,10 +2,10 @@
 {
 	internal class ItemFactory
 	{
-		public static Item GetItemById(ushort id, byte metadata)
+		public static Item GetItemById(short id, byte metadata)
 		{
 			if (id == 259) return new ItemFlintAndSteel(metadata);
-			return new Item(id, metadata);
+			return new Item((ushort) id, metadata);
 		}
 	}
 }
